pyproject.toml
[tool.poetry]
name = "shipchain-common"
version = "1.0.2"
description = "A PyPI package containing shared code for ShipChain's Python/Django projects."

license = "Apache-2.0"

authors = [
    "Adam Hodges <ahodges@shipchain.io>",
    "Clovis Djiometsa <cngnitewe@shipchain.io>",
    "James Neyer <jneyer@shipchain.io>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/ShipChain/python-common"
homepage = "https://github.com/ShipChain/python-common"

keywords = ['shipchain', 'transmission']

[tool.poetry.dependencies]
python = "^3.6"  # Compatible python versions must be declared here

aws-requests-auth = "~0.4"
boto3 = "~1.9"
cryptography = "~2.5"
django = "^2.2.4"
django-enumfields = "~0.10.0"
django-influxdb-tagged-metrics = '==1.3.3'
djangorestframework = "^3.9.1"
djangorestframework-jsonapi = "~2.5"
drf-nested-routers = "^0.91.0"
python-dateutil = "~2.7.5"
requests = "^2.21"
pytest-django = "^3.6"

[tool.poetry.dev-dependencies]
bandit = "^1.5"
coverage = "~4.5.1"
djangorestframework-simplejwt = {git = "git://github.com/ShipChain/django-rest-framework-simplejwt.git", rev = "f47065b7e0cf3adf8e4095796cfc13f651a117c6", develop = true}
prospector = "~1.1.6.2"
pytest = "^3.0"
pytest-cov = "^2.6"
pytest-mock = "^1.10"
safety = "^1.8"
